kill whole lighttpd process tree for chromium win https://bugs.webkit.org/show_bug.cgi?id=106838
Reviewed by Tony Chang.

Land a speculative fix for lighttpd.exe hanging on some chromium bots;
I think a test is causing a httpd server child process to wedge and
killing the parent httpd server process isn't sufficient to ensure
that the children are also killed.

* Scripts/webkitpy/layout_tests/servers/http_server.py:
(Lighttpd._check_and_kill):
* Scripts/webkitpy/layout_tests/servers/http_server_unittest.py:
(TestHttpServer.test_start_cmd):
(TestHttpServer):
(TestHttpServer.test_win32_start_and_stop):
(TestHttpServer.test_win32_start_and_stop.wait_for_action):
(TestHttpServer.test_win32_start_and_stop.mock_returns):
(TestHttpServer.test_win32_start_and_stop.mock_returns.return_value_thunk):


Canonical link: https://commits.webkit.org/125085@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@139697 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Tools/Scripts/webkitpy/layout_tests/servers/http_server.py
@@ -214,6 +214,15 @@ class Lighttpd(http_server_base.HttpServerBase):
 
     def _check_and_kill(self):
         if self._executive.check_running_pid(self._pid):
-            self._executive.kill_process(self._pid)
+            host = self._port_obj.host
+            if host.platform.is_win() and not host.platform.is_cygwin():
+                # FIXME: https://bugs.webkit.org/show_bug.cgi?id=106838
+                # We need to kill all of the child processes as well as the
+                # parent, so we can't use executive.kill_process().
+                #
+                # If this is actually working, we should figure out a clean API.
+                self._executive.run_command(["taskkill.exe", "/f", "/t", self._pid], error_handler=self._executive.ignore_error)
+            else:
+                self._executive.kill_process(self._pid)
             return False
         return True
